Proposal on theming `dooboo-ui` component (#377)
The `light` and the `dark` mode is default today and UI components are mandatory to handle these cases. Added ability to `theme` main components which are currently `Button` and `EditText`. We'll cover this further when this works out well.

Deprecated `@dooboo-ui/theme` and integrated into `dooboo-ui` which is the main package. This decision was made to strongly attach `theme` variables into main components.

* Release 0.0.40